SSU made non-mandatory when missing assembly

+++ scripts/ssu.bash
@@ -10,22 +10,24 @@ cd "$DIR"
 # Initialize
 date "+%Y-%m-%d %H:%M:%S %z" > "$DATASET.start"
 
-# Run barrnap
 fa="../../../05.assembly/$DATASET.LargeContigs.fna"
-barrnap --quiet --threads $CORES $fa | grep "^##gff\\|;product=16S " \
-   > $DATASET.ssu.gff 
+if [[ -s $fa ]] ; then
+   # Run barrnap
+   barrnap --quiet --threads $CORES $fa | grep "^##gff\\|;product=16S " \
+      > $DATASET.ssu.gff 
 
-# Extract
-bedtools getfasta -s "-fi" $fa -bed $DATASET.ssu.gff -fo $DATASET.ssu.all.fa 
-FastA.length.pl $DATASET.ssu.all.fa | sort -nr -k 2 | head -n 1 \
-   | cut -f 1 > $DATASET.ssu.fa.id 
-FastA.filter.pl $DATASET.ssu.fa.id $DATASET.ssu.all.fa > $DATASET.ssu.fa 
-rm $DATASET.ssu.fa.id 
-[[ -e "$fa.fai" ]] && rm "$fa.fai"
+   # Extract
+   bedtools getfasta -s "-fi" $fa -bed $DATASET.ssu.gff -fo $DATASET.ssu.all.fa 
+   FastA.length.pl $DATASET.ssu.all.fa | sort -nr -k 2 | head -n 1 \
+      | cut -f 1 > $DATASET.ssu.fa.id 
+   FastA.filter.pl $DATASET.ssu.fa.id $DATASET.ssu.all.fa > $DATASET.ssu.fa 
+   rm $DATASET.ssu.fa.id 
+   [[ -e "$fa.fai" ]] && rm "$fa.fai"
 
-# Gzip
-gzip -9 -f "$DATASET.ssu.gff"
-gzip -9 -f "$DATASET.ssu.all.fa"
+   # Gzip
+   gzip -9 -f "$DATASET.ssu.gff"
+   gzip -9 -f "$DATASET.ssu.all.fa"
+fi
 
 # Finalize
 date "+%Y-%m-%d %H:%M:%S %z" > "$DATASET.done"
